Users can now make their sessions public for others to easily find.
